move processSynapseJSON to different file, preparing to use it in script
